requirements.txt
Django
djangorestframework
python-dotenv
djoser
django-cors-headers
djangorestframework-simplejwt
django-rest-swagger
django-filter
django-model-utils